requirements.txt
pandas 
numpy 
scikit-learn 
lightgbm 
matplotlib 
seaborn
imblearn
pdpbox
python-dotenv
joblib
lightgbm
json